refactor: make useShouldShowReactionModal and useShouldShowAchievementSuccessModal return a modal state
